backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0

# Supabase (let it resolve its own dependencies)
supabase

# File uploads
python-multipart==0.0.6

# Date handling
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
